\name{Hmat}
\alias{Hmat}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Matrix algebra transformation}
\description{
Internal function that performs a matrix algebra transformation that is used for estimation in \code{\link{extRC}}.
}
\usage{
Hmat(G)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{G}{input matrix}
}

\value{
\item{H}{transformed matrix}
}

\author{Francesco Bartolucci, Antonio Forcina
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{algebra}% use one of  RShowDoc("KEYWORDS")
