\name{mobility}
\alias{mobility}
\docType{data}
\title{Social mobility data}
\description{
Social mobility table of 3,500 British individuals, who are cross-classified according to their occupational status and the occupation status of their fathers.
}
\usage{data("mobility")}
\format{
  The format is:
 num [1:5, 1:5] 50 28 11 14 3 45 174 78 150 42 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:5] "F1" "F2" "F3" "F4" ...
  ..$ : chr [1:5] "S1" "S2" "S3" "S4" ...
}

\references{
Mosteller, F. (1968). Association and estimation in contingency tables. \emph{Journal of the American Statistical Association}, \bold{63}, 1-28.

Bartolucci, F. and Forcina, A. (2002). Extended RC association models allowing for order restrictions and marginal modeling. \emph{Journal of the American Statistical Association}, \bold{97}, 1192-1199.
}

\keyword{datasets}
