\name{plot}
\alias{plot.extRC}

\title{Plot the output}
\description{It plots the output of code{\link{extRC}} function for a vector of lambda values.}

\usage{
\method{plot}{extRC}(x, ...)
}


\arguments{
\item{x}{output from \code{\link{extRC}}}
\item{...}{further arguments passed to or from other methods}
}

\value{
\item{None}{}
}

\author{Francesco Bartolucci, Antonio Forcina}
