\name{print}
\alias{print.extRC}

\title{Print the output.}
\description{Given the output of code{\link{extRC}} function, it is written in a readable form.}

\usage{
\method{print}{extRC}(x, ...)
}


\arguments{
\item{x}{output from \code{\link{extRC}}}
\item{...}{further arguments passed to or from other methods}
}

\value{
\item{None}{}
}

\author{Francesco Bartolucci, Antonio Forcina
}
