\name{summary}
\alias{summary.extRC}

\title{Summary of extRC fits}
\description{Summary method for the output of code{\link{extRC}} function.}

\usage{
\method{summary}{extRC}(object, ...)
}


\arguments{
\item{object}{output from \code{\link{extRC}}}
\item{...}{further arguments passed to or from other methods}
}

\value{
\item{None}{}
}

\author{Francesco Bartolucci, Antonio Forcina}